+++ src/main/java/org/broadinstitute/hellbender/utils/gvs/bigquery/PendingBQWriter.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.utils.bigquery;
+package org.broadinstitute.hellbender.utils.gvs.bigquery;
 
 import com.google.cloud.bigquery.storage.v1beta2.*;
 import org.apache.logging.log4j.LogManager;
