+++ src/main/java/org/broadinstitute/hellbender/utils/gvs/localsort/AvroSortingCollectionCodec.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.utils.localsort;
+package org.broadinstitute.hellbender.utils.gvs.localsort;
 
 import org.apache.avro.Schema;
 import org.apache.avro.file.DataFileStream;
