+++ src/main/java/org/broadinstitute/hellbender/utils/gvs/localsort/SortingCollection.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.utils.localsort;
+package org.broadinstitute.hellbender.utils.gvs.localsort;
 
 import htsjdk.samtools.Defaults;
 import htsjdk.samtools.util.*;
